bugfix totalDiscounted floating division

+++ features/invoices/data/repositories/updateInvoiceComponent.go
@@ -236,7 +236,7 @@ func (updateInvoiceTrx *updateInvoiceTransactionComponent) TransactionBody(
 	}
 
 	if discountInPercent > 0 {
-		totalDiscounted = int(math.Round(float64(discountInPercent / 100.0 * totalPrice)))
+		totalDiscounted = int(math.Round(float64(discountInPercent) / 100.0 * float64(totalPrice)))
 	}
 	invoiceToUpdate.TotalDiscounted = &totalDiscounted
 	invoiceToUpdate.TotalPayable = func(i int) *int { return &i }(totalPrice - totalDiscounted)
